+++ Model/CBetting.cs
@@ -20,10 +20,10 @@ namespace LSportsServer
             m_nMemberSn = CGlobal.ParseInt(info["member_sn"]);
             m_strBettingNo = Convert.ToString(info["betting_no"]);
             m_nSelectNo = CGlobal.ParseInt(info["select_no"]);
-            m_fHomeRate = Convert.ToSingle(info["home_rate"]);
-            m_fAwayRate = Convert.ToSingle(info["away_rate"]);
-            m_fDrawRate = Convert.ToSingle(info["draw_rate"]);
-            m_fSelectRate = Convert.ToSingle(info["select_rate"]);
+            m_fHomeRate = Convert.ToDouble(info["home_rate"]);
+            m_fAwayRate = Convert.ToDouble(info["away_rate"]);
+            m_fDrawRate = Convert.ToDouble(info["draw_rate"]);
+            m_fSelectRate = Convert.ToDouble(info["select_rate"]);
             m_strBetID = Convert.ToString(info["betid"]);
             m_nGameType = CGlobal.ParseInt(info["game_type"]);
             m_nResult = CGlobal.ParseInt(info["result"]);
